+++ backend/src/main/java/tuneandmanner/wiselydiarybackend/diary/service/DiaryService.java
@@ -9,9 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import tuneandmanner.wiselydiarybackend.diary.domain.entity.Diary;
-import tuneandmanner.wiselydiarybackend.diary.domain.entity.DiarySummary;
+import tuneandmanner.wiselydiarybackend.diarysummary.domain.DiarySummary;
 import tuneandmanner.wiselydiarybackend.diary.domain.repository.DiaryRepository;
-import tuneandmanner.wiselydiarybackend.diary.domain.repository.DiarySummaryRepository;
+import tuneandmanner.wiselydiarybackend.diarysummary.repository.DiarySummaryRepository;
 import tuneandmanner.wiselydiarybackend.rag.service.RAGService;
 
 import java.util.Map;
