+++ backend/src/main/java/tuneandmanner/wiselydiarybackend/diarysummary/repository/DiarySummaryRepository.java
@@ -1,0 +1,15 @@
+package tuneandmanner.wiselydiarybackend.diarysummary.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import tuneandmanner.wiselydiarybackend.diarysummary.domain.DiarySummary;
+
+@Repository
+public interface DiarySummaryRepository extends JpaRepository<DiarySummary, Long> {
+	DiarySummary findByDiaryCode(Long diaryCode);
+	Optional<DiarySummary> findByDiarySummaryCode(@Param("diarySummaryCode") Long diarySummaryCode);
+}
