+++ backend/src/main/java/tuneandmanner/wiselydiarybackend/letter/service/LetterService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tuneandmanner.wiselydiarybackend.common.exception.NotFoundException;
-import tuneandmanner.wiselydiarybackend.diary.domain.entity.DiarySummary;
-import tuneandmanner.wiselydiarybackend.diary.domain.repository.DiarySummaryRepository;
+import tuneandmanner.wiselydiarybackend.diarysummary.domain.DiarySummary;
+import tuneandmanner.wiselydiarybackend.diarysummary.repository.DiarySummaryRepository;
 import tuneandmanner.wiselydiarybackend.letter.domain.entity.Letter;
 import tuneandmanner.wiselydiarybackend.letter.domain.repository.LetterRepository;
 import tuneandmanner.wiselydiarybackend.letter.dto.response.CreateLetterResponse;
